+++ changelog.md
@@ -12,4 +12,5 @@
 
 # Next
 
+* Fix: `.clone()` with `subclassZipFile` option breaks `lazyEntries: false` (closes #1)
 * Remove unnecessary instanceof checks in patchers